chore: rework file system
