+++ lib/src/components/Modal/Assets/AssetBackdrop.tsx
