+++ lib/src/components/Modal/Assets/AssetContent.tsx
