+++ lib/src/components/Modal/Assets/CloseButton.tsx
