+++ lib/src/components/Modal/Assets/assets.module.scss
