--- lib/src/components/Modal/components/Close/types.ts
@@ -1,6 +1,0 @@
-//TODO Migrate CloseButton, but to what?
-import type { CloseButtonProps } from '@old/CloseButton'
-
-export type CloseProps = CloseButtonProps & {
-  isOnHeader?: boolean
-}
