+++ lib/src/components/Modal/components/Cover.tsx
@@ -1,0 +1,15 @@
+import { useTheme } from 'styled-components'
+
+//TODO Migrate shape, but to what?
+import type { ShapeProps } from '@old/Shape'
+import { Shape } from '@old/Shape'
+
+export const Cover: React.FC<ShapeProps> = props => {
+  const { modals } = useTheme()
+
+  return (
+    <div>
+      <Shape {...modals.cover} {...props} />
+    </div>
+  )
+}
