+++ lib/src/components/Modal/components/Footer.tsx
@@ -1,10 +1,9 @@
 import { forwardRef } from 'react'
 
+import modalStyles from '@/components/Modal/modal.module.scss'
+import type { FooterProps } from '@/components/Modal/types'
 import { Text } from '@/components/Text'
 import { classNames } from '@/utils/classNames'
-
-import modalStyles from './footer.module.scss'
-import type { FooterProps } from './types'
 
 const cx = classNames(modalStyles)
