+++ lib/src/components/Modal/modal.module.scss
@@ -81,10 +81,57 @@
     }
   }
 
+  .header {
+    position: sticky;
+    top: 0;
+    z-index: 1;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: var(--color-neutral-10);
+    border-top-left-radius: var(--radius-md);
+    border-top-right-radius: var(--radius-md);
+    padding-bottom: var(--spacing-xxl);
+    padding-left: var(--spacing-xxl);
+    padding-right: var(--spacing-3xl);
+    padding-top: var(--spacing-xxl);
+  }
+
+  .header-subtitle {
+    color: var(--color-neutral-70);
+    margin: 0;
+  }
+
   .body {
     color: var(--color-neutral-90);
     padding-left: var(--spacing-xxl);
     padding-right: var(--spacing-3xl);
     padding-bottom: var(--spacing-xxl);
+  }
+
+  .footer {
+    position: sticky;
+    bottom: 0;
+    width: 100%;
+    flex-shrink: 0;
+    z-index: 1;
+    background-color: var(--color-neutral-10);
+    border-bottom-left-radius: var(--radius-md);
+    border-bottom-right-radius: var(--radius-md);
+    border-top: 0 solid var(--color-neutral-30);
+  }
+
+  .footer-children-wrapper {
+    padding-block: var(--spacing-xl);
+    padding-inline: var(--spacing-xxl);
+  }
+
+  .footer-information {
+    background-color: var(--color-beige-20);
+    padding-bottom: var(--spacing-xxl);
+    padding-left: var(--spacing-xxl);
+    padding-right: var(--spacing-xxl);
+    padding-top: var(--spacing-xl);
   }
 }
